+++ argopy/data_fetchers/argovis_data.py
@@ -15,6 +15,7 @@ import getpass
 from .proto import ArgoDataFetcherProto
 from abc import abstractmethod
 import warnings
+import urllib
 
 from argopy.stores import httpstore
 from argopy.options import OPTIONS
@@ -135,6 +136,106 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
             this.attrs["history"] = txt
         return this
 
+    def _add_attributes(self, this):  # noqa: C901
+        """ Add variables attributes not return by erddap requests (csv)
+
+            This is hard coded, but should be retrieved from an API somewhere
+        """
+        for v in this.data_vars:
+            if "TEMP" in v and "_QC" not in v:
+                this[v].attrs = {
+                    "long_name": "SEA TEMPERATURE IN SITU ITS-90 SCALE",
+                    "standard_name": "sea_water_temperature",
+                    "units": "degree_Celsius",
+                    "valid_min": -2.0,
+                    "valid_max": 40.0,
+                    "resolution": 0.001,
+                }
+                if "ERROR" in v:
+                    this[v].attrs["long_name"] = (
+                        "ERROR IN %s" % this[v].attrs["long_name"]
+                    )
+
+        for v in this.data_vars:
+            if "PSAL" in v and "_QC" not in v:
+                this[v].attrs = {
+                    "long_name": "PRACTICAL SALINITY",
+                    "standard_name": "sea_water_salinity",
+                    "units": "psu",
+                    "valid_min": 0.0,
+                    "valid_max": 43.0,
+                    "resolution": 0.001,
+                }
+                if "ERROR" in v:
+                    this[v].attrs["long_name"] = (
+                        "ERROR IN %s" % this[v].attrs["long_name"]
+                    )
+
+        for v in this.data_vars:
+            if "PRES" in v and "_QC" not in v:
+                this[v].attrs = {
+                    "long_name": "Sea Pressure",
+                    "standard_name": "sea_water_pressure",
+                    "units": "decibar",
+                    "valid_min": 0.0,
+                    "valid_max": 12000.0,
+                    "resolution": 0.1,
+                    "axis": "Z",
+                }
+                if "ERROR" in v:
+                    this[v].attrs["long_name"] = (
+                        "ERROR IN %s" % this[v].attrs["long_name"]
+                    )
+
+        for v in this.data_vars:
+            if "DOXY" in v and "_QC" not in v:
+                this[v].attrs = {
+                    "long_name": "Dissolved oxygen",
+                    "standard_name": "moles_of_oxygen_per_unit_mass_in_sea_water",
+                    "units": "micromole/kg",
+                    "valid_min": -5.0,
+                    "valid_max": 600.0,
+                    "resolution": 0.001,
+                }
+                if "ERROR" in v:
+                    this[v].attrs["long_name"] = (
+                        "ERROR IN %s" % this[v].attrs["long_name"]
+                    )
+
+        for v in this.data_vars:
+            if "_QC" in v:
+                attrs = {
+                    "long_name": "Global quality flag of %s profile" % v,
+                    "convention": "Argo reference table 2a",
+                }
+                this[v].attrs = attrs
+
+        if "CYCLE_NUMBER" in this.data_vars:
+            this["CYCLE_NUMBER"].attrs = {
+                "long_name": "Float cycle number",
+                "convention": "0..N, 0 : launch cycle (if exists), 1 : first complete cycle",
+            }
+
+        if "DATA_MODE" in this.data_vars:
+            this["DATA_MODE"].attrs = {
+                "long_name": "Delayed mode or real time data",
+                "convention": "R : real time; D : delayed mode; A : real time with adjustment",
+            }
+
+        if "DIRECTION" in this.data_vars:
+            this["DIRECTION"].attrs = {
+                "long_name": "Direction of the station profiles",
+                "convention": "A: ascending profiles, D: descending profiles",
+            }
+
+        if "PLATFORM_NUMBER" in this.data_vars:
+            this["PLATFORM_NUMBER"].attrs = {
+                "long_name": "Float unique identifier",
+                "convention": "WMO float identifier : A9IIIII",
+            }
+
+        return this
+
     @property
     def cachepath(self):
         """ Return path to cache file for this request """
@@ -143,6 +244,15 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
     def cname(self):
         """ Return a unique string defining the constraints """
         return self._cname()
+
+    def url_encode(self, urls):
+        """ Return safely encoded list of urls
+
+            This was made to debug for fsspec caching system not working with cache of profile and region in argovis
+            Not working yet, see: https://github.com/euroargodev/argopy/issues/101
+        """
+        return urls
+        # return [urllib.parse.quote(url, safe='/:?=[]&') for url in urls]
 
     def json2dataframe(self, profiles):
         """ convert json data to Pandas DataFrame """
@@ -206,7 +316,9 @@ class ArgovisDataFetcher(ArgoDataFetcherProto):
         ds = ds.set_coords(coords)
 
         # Cast data types and add variable attributes (not available in the csv download):
+        ds['TIME'] = ds['TIME'].astype(np.datetime64)
         ds = ds.argo.cast_types()
+        ds = self._add_attributes(ds)
 
         # Remove argovis file attributes and replace them with argopy ones:
         ds.attrs = {}
@@ -318,7 +430,8 @@ class Fetch_wmo(ArgovisDataFetcher):
                     this.append(self.get_url(wmo, cycs))
             return this
 
-        return list_bunch(self.WMO, self.CYC)
+        urls = list_bunch(self.WMO, self.CYC)
+        return self.url_encode(urls)
 
     def dashboard(self, **kw):
         if len(self.WMO) == 1:
@@ -389,8 +502,8 @@ class Fetch_box(ArgovisDataFetcher):
         return url
 
     def get_url(self):
-        # return self.get_url_shape()
-        return self.get_url_rect()
+        return self.get_url_shape()
+        # return self.get_url_rect()
 
     @property
     def uri(self):
@@ -404,6 +517,7 @@ class Fetch_box(ArgovisDataFetcher):
         MaxLenTime = 90
         MaxLen = np.timedelta64(MaxLenTime, "D")
 
+        urls = []
         if not self.parallel:
             # Check if the time range is not larger than allowed (90 days):
             if Lt > MaxLen:
@@ -413,12 +527,10 @@ class Fetch_box(ArgovisDataFetcher):
                     chunksize={"time": MaxLenTime},
                 )
                 boxes = self.Chunker.fit_transform()
-                urls = []
                 for box in boxes:
                     urls.append(Fetch_box(box=box, ds=self.dataset_id).get_url())
-                return urls
             else:
-                return [self.get_url()]
+                urls.append(self.get_url())
         else:
             if 'time' not in self.chunks_maxsize:
                 self.chunks_maxsize['time'] = MaxLenTime
@@ -435,12 +547,12 @@ class Fetch_box(ArgovisDataFetcher):
                 {"box": self.BOX}, chunks=self.chunks, chunksize=self.chunks_maxsize
             )
             boxes = self.Chunker.fit_transform()
-            urls = []
             for box in boxes:
                 urls.append(Fetch_box(box=box, ds=self.dataset_id).get_url())
-            return urls
+
+        return self.url_encode(urls)
 
     @property
     def url(self):
-        # return self.get_url_shape()
-        return self.get_url_rect()
+        return self.get_url_shape()
+        # return self.get_url_rect()